fix youtube embed, remove activityInfo.jsx
