+++ package.json
@@ -10,7 +10,8 @@
   },
   "dependencies": {
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-youtube": "^9.0.3"
   },
   "devDependencies": {
     "@types/react": "^18.0.15",
