+++ src/components/ActivityCard/ActivityCard.css
@@ -1,4 +1,5 @@
 .activity-card {
+     width: 250px;
      border: #50a3a2 1px solid;
      font-family: "Source Sans Pro", sans-serif;
      padding:15px ;
@@ -8,6 +9,7 @@
      align-items: center;
      box-shadow: 5px 5px 5px rgb(132, 132, 132);
      gap: 10px 20px;
+     overflow: hidden;
 }
 
 .activity-card img {
@@ -16,7 +18,7 @@
 }
 
 .activity-card video {
-    height: 100%;
+    
     border-radius: 10px;
     border-radius: 15px;
     border: solid 2px red;
