--- src/components/ActivityCardList/ActivityInfo.jsx
@@ -1,28 +1,0 @@
-const activityInfo = [
-    {
-        id: 1,
-        img: "../src/assets/imageCard.png",
-        video: "https://www.youtube.com/watch?v=0J7D21OEbC4",
-        activity: "yoca",
-        durationTime: "3hr 5s",
-       
-    },
-    {
-        id: 2,
-        img: "../src/assets/yoca.jpg",
-        video: "https://youtu.be/0J7D21OEbC4",
-        activity: "yoca2",
-        durationTime: "1hr 5s",
-     
-    },
-    {
-        id: 3,
-        img: "../src/assets/run.jpg",
-        video: "https://www.youtube.com/embed/yMNLa17g9uc",
-        activity: "yoca3",
-        durationTime: "2hr 5s",
-       
-    }
-]
-
-export default activityInfo;
